refactor: divide legacy jdbc template code
